First "playable" version and fixes
The game Aces to Kings is now playable though only as a single player,
and there's no reasonable way for the player to know what's going on
either (but it's playable!). Also lots of small fixes to avoid compiler
errors and a couple of fixes for how the game should actually play.