Progress towards a better choice system
Created the Selector class with the single generic method select. This
allows for removing a lot of the previous redundancy in PlayerChoice and
Turn. Also allows for displaying the name of cards when selecting from a
list, rather than simply CARD_0, CARD_1, etc. as was previously the
case. As a final note, some smaller fixes were made to round.