Refactoring of melds
Essentially an overhaul of how the two meld classes previously
functioned. Added an abstract class from which both kinds of melds are
extended, with a single public method play(player, cards). This method
will return true if the cards given have been successfully played and
false if not.

Within the meld classes themselves, a great deal of the verification
process for whether cards can be played has been refactored and
simplified to enhance the readability of the classes. As a result of
this, a couple of bugs were fixed in the process.

This will break the Turn class and consequently the game is currently
unplayable. A similar refactoring of Turn is to come. Additionally,
CardBank names have been removed from Round and placed into their own
class CardBanks.