Card package rework
A large rework of the cardgame.card package has been performed. This
will break the game for the time being. The following refers exclusively
to the cardgame.card package.

A large amount of documenting has been completed.

CardBank has been renamed to Bank. Lots of methods have been
removed/rewritten to avoid redundancy and hopefully reduce coupling.

Deck no longer extends Bank. The class itself has been rewritten to
provide more specific functionality for decks of cards.

A new interface CardCollection has been created. This is implemented by
Bank and Deck. It is planned for the melds in aces to kings to implement
this eventually. This interface contains basic methods surrounding
collections of cards.

A new interface Drawable has been created, which extends CardCollection.
This interface is implemented by Bank and Deck. This interface intends
to provide methods for drawing cards from a CardCollection.

A new value has been added to the Rank and Suit enums - JOKER. This is
intended to provide special functionality for cards when it is needed.

The Suit enum now also provides details about the colours of each suit.

The compareTo method in Card has been simplified.

The cardgame.events package has been deleted. The functionality of the
classes previously there has been migrated over to the cardgame.card
package.

The scanner in ConsolePlayerIO in cardgame.player is now properly closed
after use.