Melds package rework
Changes outside of the melds package are as follows:
Updated the README.

Added remove and default transferTo to the CardCollection interface, and
implemented the remove method in implementing classes.

Added getNeighbour to the Rank enum.

Changes to the melds package are as follows:
All classes have been documented to some extent.

AbstractMeld now implement the CardCollection interface. Extensions of
AbstractMeld now contain the implemented methods.

Removed the constructor from AbstractMeld. Melds will no longer need to
keep track of the joker rank in the next commit (which should make the
game playable again).

MeldsManager now has a reset method to reset all of its meld to empty.
References to melds now deal with the AbstractMeld class rather than the
extensions.

RankMeld now maintains a list of jokers rather than counting them and
inserting them to the front. Many methods have received changes to
properly implement the CardCollection interface. Various other small
changes have also been made.

The RunMeld class has been almost completely rewritten. Aces can now be
played as a high or a low card, and not both at the same time. The size
of the class has grown tremendously and it may be a candidate for
refactoring in a later commit.

PlayOption has received some extra functionality to support RankMeld.
I'm not too happy with the implementation of this class and will revisit
this in the future.