Restructured card interfaces
The Card class has been renamed PlayingCard. The compareTo method has
been simplified again. PlayingCard now implements a new abstract class
Card.

The CardCollection interface now extends Selectable as opposed to
Drawable extending Selectable. The default transferTo method now throws
an exception if a specified card to remove does not exist in the
collection. This interface now implements generics in referring to
classes extending the Card class.

The Bank class has been split into two classes, the first of which is
called Hand. This class implements the CardCollection interface and
provides methods specific to collections of cards owned by players.

The Drawable interface now specifies the methods previously in the
DeckEventSource interface, which has been removed. The DeckEventListener
interface has been renamed to DrawListener.

The second half of the previous Bank class is now implemented as a new
class Stack. This class implements Drawable and provides methods for a
last-in-first-out stack of cards.

The Deck class has been generalised to accept any type of Card.
Construction of decks now requires a list of initial cards.