Created AcesToKingsDeck
Created AcesToKingsDeck. This provides more specialised functionality
for the game Aces to Kings, such as creating a swapping out cards in the
deck for jokers between rounds. The game has been updated to use this
new adapted deck.

The Deck class now requires an int argument in the constructor, which
will specify how many cards to deal when the new Deal(CardCollection)
method is called.

The sole select method in Selector now requires a generic list as
opposed to a generic array. Methods elsewhere which called select have
been updated accordingly.

A few minor fixes have been applied in the rest of the changed classes.