Iterable cards and shifting some methods
The Drawable interface now also extends Iterable. Hand also implements
Iterable. This allows for iterating over the cards in these collections
as you would a list. This has allowed for some simplification of methods
elsewhere.

The chooseCards method in Turn has been removed and a more general
method has been added to the Hand class, providing the same
functionality.

A bug was fixed in the RunMeld class that allowed joker cards to be
added to any empty slot, so long as a meld existed.